Moved extraction and q_metric logic to cpp (cherry picked from commit 166aa0d)
